+++ system/src/Grav/Common/Page/Types.php
@@ -33,7 +33,7 @@ class Types implements \ArrayAccess, \Iterator, \Countable
         }
 
         if (!$blueprint && $this->systemBlueprints) {
-            $blueprint = isset($this->systemBlueprints[$type]) ? $this->systemBlueprints[$type] : $this->systemBlueprints['default'];
+            $blueprint = $this->systemBlueprints[$type] ?? $this->systemBlueprints['default'];
         }
 
         if ($blueprint) {
@@ -43,7 +43,7 @@ class Types implements \ArrayAccess, \Iterator, \Countable
 
     public function scanBlueprints($uri)
     {
-        if (!is_string($uri)) {
+        if (!\is_string($uri)) {
             throw new \InvalidArgumentException('First parameter must be URI');
         }
 
@@ -63,7 +63,7 @@ class Types implements \ArrayAccess, \Iterator, \Countable
 
     public function scanTemplates($uri)
     {
-        if (!is_string($uri)) {
+        if (!\is_string($uri)) {
             throw new \InvalidArgumentException('First parameter must be URI');
         }
 
@@ -96,7 +96,7 @@ class Types implements \ArrayAccess, \Iterator, \Countable
             if (strpos($name, '/')) {
                 continue;
             }
-            $list[$name] = ucfirst(strtr($name, '_', ' '));
+            $list[$name] = ucfirst(str_replace('_', ' ', $name));
         }
         ksort($list);
         return $list;
@@ -109,7 +109,7 @@ class Types implements \ArrayAccess, \Iterator, \Countable
             if (strpos($name, 'modular/') !== 0) {
                 continue;
             }
-            $list[$name] = trim(ucfirst(strtr(basename($name), '_', ' ')));
+            $list[$name] = ucfirst(trim(str_replace('_', ' ', basename($name))));
         }
         ksort($list);
         return $list;